# RGB to YUV Performance Improvements