Changed to use integer math instead of f32 method. Adjusted to use iterators instead of array indexing.

New benchmarking timings on my PC (Windows 10, AMD Ryzen 7900X)

```
test convert_rgb8_to_yuv_1920x1080 ... bench:     305,200.00 ns/iter (+/- 26,439.00)
test convert_rgb8_to_yuv_512x512   ... bench:      44,606.80 ns/iter (+/- 1,529.76)
```

Old benchmarking timings
```
test convert_rgb8_to_yuv_1920x1080 ... bench:   4,473,440.00 ns/iter (+/- 368,726.00)
test convert_rgb8_to_yuv_512x512   ... bench:     543,745.00 ns/iter (+/- 9,916.00)
```