@ralfbiedert 2025-09-05
BTW, I can confirm a ~14x performance increase here locally, absolutely outstanding!